request format issue in swagger

+++ src/main/java/com/tm/evaluationsystem/model/Result.java
@@ -22,6 +22,6 @@ public class Result {
     private List<QuestionAnswerMap> questionAnswerMaps;
     private int attemptedQuestions;
     private int rightAnswers;
-    private int percentage;
+    private float percentage;
     private boolean isPassed;
 }
